# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PriceCheck NG API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors/{id}/reviews:
    get:
      operationId: listVendorReviews
      tags: [reviews]
      summary: List reviews for a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorReview"
    post:
      operationId: createVendorReview
      tags: [reviews]
      summary: Submit a review for a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorReview"
...
  /prices:
    get:
      operationId: listPrices
      tags: [prices]
      summary: Compare prices for a product across vendors
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: integer
        - name: location
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [low_to_high, high_to_low, nearest, recently_updated, best_rated]
      responses:
        "200":
          description: Price comparison results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceEntry"
    post:
      operationId: submitPrice
      tags: [prices]
      summary: Submit a price for a product at a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceInput"
      responses:
        "201":
          description: Price submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceEntry"
  /prices/{id}:
    patch:
      operationId: updatePrice
      tags: [prices]
      summary: Update a price listing (vendor dashboard)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceUpdateInput"
      responses:
        "200":
          description: Updated price entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceEntry"
  /price-history:
    get:
      operationId: getPriceHistory
      tags: [prices]
      summary: Get 30-day price history for a product
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: integer
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceHistoryPoint"
  /price-summary:
    get:
      operationId: getPriceSummary
      tags: [prices]
      summary: Get price summary stats for a product (avg, min, max, trend)
      parameters:
        - name: product_id
          in: query
          required: true
          schema:
            type: integer
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Price summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceSummary"
...
components:
  schemas:
...
    VendorReview:
      type: object
      required: [id, vendor_id, rating, reviewer_name, created_at]
      properties:
        id:
          type: integer
        vendor_id:
          type: integer
        rating:
          type: integer
        reviewer_name:
          type: string
        comment:
          type: ["string", "null"]
        created_at:
          type: string
    VendorReviewInput:
      type: object
      required: [rating, reviewer_name]
      properties:
        rating:
          type: integer
        reviewer_name:
          type: string
        comment:
          type: string
    PriceEntry:
      type: object
      required: [id, product_id, vendor_id, price, quantity, updated_at]
      properties:
        id:
          type: integer
        product_id:
          type: integer
        vendor_id:
          type: integer
        vendor_name:
          type: string
        vendor_location:
          type: string
        vendor_logo_url:
          type: ["string", "null"]
        vendor_verified:
          type: boolean
        vendor_whatsapp:
          type: ["string", "null"]
        vendor_rating:
          type: ["number", "null"]
        vendor_rating_count:
          type: ["integer", "null"]
        vendor_stock_available:
          type: boolean
        vendor_delivery_options:
          type: ["string", "null"]
        vendor_response_time:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: string
        is_cheapest:
          type: boolean
        is_best_rated:
          type: boolean
        updated_at:
          type: string
    PriceInput:
      type: object
      required: [product_id, vendor_id, price, quantity]
      properties:
        product_id:
          type: integer
        vendor_id:
          type: integer
        price:
          type: number
        quantity:
          type: string
    PriceUpdateInput:
      type: object
      properties:
        price:
          type: number
        quantity:
          type: string
        stock_available:
          type: boolean
    PriceHistoryPoint:
      type: object
      required: [date, average_price, min_price]
      properties:
        date:
          type: string
        average_price:
          type: number
        min_price:
          type: number
    PriceSummary:
      type: object
      required: [product_id, average_price, min_price, max_price, price_change_pct, vendor_count]
      properties:
        product_id:
          type: integer
        average_price:
          type: number
        min_price:
          type: number
        max_price:
          type: number
        price_change_pct:
          type: number
        vendor_count:
          type: integer
        cheapest_vendor:
          type: ["string", "null"]
...