# lib/api-zod/src/generated/types/priceEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceEntry {
  id: number;
  product_id: number;
  vendor_id: number;
  vendor_name?: string;
  vendor_location?: string;
  /** @nullable */
  vendor_logo_url?: string | null;
  vendor_verified?: boolean;
  /** @nullable */
  vendor_whatsapp?: string | null;
  /** @nullable */
  vendor_rating?: number | null;
  /** @nullable */
  vendor_rating_count?: number | null;
  vendor_stock_available?: boolean;
  /** @nullable */
  vendor_delivery_options?: string | null;
  /** @nullable */
  vendor_response_time?: string | null;
  price: number;
  quantity: string;
  is_cheapest?: boolean;
  is_best_rated?: boolean;
  updated_at: string;
}