# lib/api-zod/src/generated/types/priceUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceUpdateInput {
  price?: number;
  quantity?: string;
  stock_available?: boolean;
}