# lib/api-zod/src/generated/types/vendorReviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PriceCheck NG API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VendorReviewInput {
  rating: number;
  reviewer_name: string;
  comment?: string;
}